+++ src/com/lapissea/cfs/objects/text/Encoding.java
@@ -87,7 +87,7 @@ class Encoding{
 			try(var stream=new BitInputStream(w)){
 				for(int i=0;i<text.charCount();i++){
 					stream.prepareBits(bits);
-					int  index=stream.readBits(bits);
+					int  index=(int)stream.readBits(bits);
 					char c    =decode(index);
 					
 					sb.append(c);
